refactor(home/setup): bring implementation in same file of the traits
