+++ src/main/scala/scatan/Main.scala
@@ -1,12 +1,12 @@
 package scatan
 import com.raquo.laminar.api.L.{*, given}
-import scatan.example.controller.{AboutController, AboutControllerImpl}
 import scatan.controllers.home.{HomeController, HomeControllerImpl}
 import scatan.controllers.setup.{SetUpController, SetUpControllerImpl}
 import scatan.views.setup.{SetUpView, ScalaJsSetUpView}
 import scatan.views.home.{HomeView, ScalaJsHomeView}
-import scatan.example.model.CounterAppState
-import scatan.example.view.{AboutView, ScalaJSAboutView}
+import scatan.model.ApplicationState
+import scatan.views.home.{AboutView, ScalaJSAboutView}
+import scatan.controllers.home.{AboutController, AboutControllerImpl}
 import scatan.mvc.lib.application.NavigableApplication
 import scatan.mvc.lib.page.PageFactory
 import scatan.mvc.lib.{Controller, Model, NavigableApplicationManager, ScalaJSView}
@@ -14,34 +14,34 @@ import scatan.mvc.lib.{Controller, Model, NavigableApplicationManager, ScalaJSVi
 import scala.util.Random
 
 // Route
-enum Pages(val pageFactory: PageFactory[?, ?, CounterAppState]):
+enum Pages(val pageFactory: PageFactory[?, ?, ApplicationState]):
   case Home
       extends Pages(
-        PageFactory[HomeController, HomeView, CounterAppState](
+        PageFactory[HomeController, HomeView, ApplicationState](
           viewFactory = new ScalaJsHomeView(_, "root"),
           controllerFactory = new HomeControllerImpl(_)
         )
       )
   case Setup
       extends Pages(
-        PageFactory[SetUpController, SetUpView, CounterAppState](
+        PageFactory[SetUpController, SetUpView, ApplicationState](
           viewFactory = new ScalaJsSetUpView(_, "root"),
           controllerFactory = new SetUpControllerImpl(_)
         )
       )
   case About
       extends Pages(
-        PageFactory[AboutController, AboutView, CounterAppState](
+        PageFactory[AboutController, AboutView, ApplicationState](
           viewFactory = new ScalaJSAboutView(_, "root"),
           controllerFactory = new AboutControllerImpl(_)
         )
       )
 
-// Application
-val CounterApplication: NavigableApplication[CounterAppState, Pages] = NavigableApplication[CounterAppState, Pages](
-  initialState = CounterAppState(0),
+// App
+val Application: NavigableApplication[ApplicationState, Pages] = NavigableApplication[ApplicationState, Pages](
+  initialState = ApplicationState(0),
   pagesFactories = Pages.values.map(p => p -> p.pageFactory).toMap
 )
 
 @main def main(): Unit =
-  NavigableApplicationManager.startApplication(CounterApplication, Pages.Home)
+  NavigableApplicationManager.startApplication(Application, Pages.Home)
