+++ src/main/scala/scatan/model/ApplicationModel.scala
@@ -1,0 +1,5 @@
+package scatan.model
+
+import scatan.mvc.lib.Model
+
+final case class ApplicationState(boo: Int) extends Model.State
